updated breeze plugin

+++ wp-content/plugins/breeze/inc/cache/purge-cache.php
@@ -24,10 +24,16 @@ if ( ! defined( 'ABSPATH' ) ) {
 
 class Breeze_PurgeCache {
 
+	/**
+	 * Temporary page/post/cpt ID list of already cache cleared.
+	 */
+	private array $cleared_extra_items = array();
+
 	public function set_action() {
 		add_action( 'pre_post_update', array( $this, 'purge_post_on_update' ), 10, 1 );
 		add_action( 'save_post', array( $this, 'purge_post_on_update' ), 10, 1 );
 		add_action( 'save_post', array( $this, 'purge_post_on_update_content' ), 9, 3 );
+
 		add_action( 'edited_term', array( $this, 'purge_term_on_update' ), 9, 3 );
 		add_action( 'wp_trash_post', array( $this, 'purge_post_on_update' ), 10, 1 );
 		add_action( 'wp_trash_post', array( $this, 'purge_post_on_trash' ), 9, 1 );
@@ -40,6 +46,54 @@ class Breeze_PurgeCache {
 
 		add_action( 'switch_theme', array( &$this, 'clear_local_cache_on_switch' ), 9, 3 );
 		add_action( 'customize_save_after', array( &$this, 'clear_customizer_cache' ), 11, 1 );
+	}
+
+	/**
+	 * Detects pages with the latest comments block and clears their cache.
+	 *
+	 * This method scans the content of published posts to find pages that contain
+	 * the latest comments block (`<!-- wp:latest-comments -->`). If such pages are found,
+	 * it purges their Cloudflare cache and removes local cache files.
+	 *
+	 * @return void
+	 */
+	private function detect_comments_page_clear_cache(): void {
+		global $wpdb;
+
+		$query = "SELECT ID
+			FROM `$wpdb->posts`
+			WHERE post_content LIKE '%<!-- wp:latest-comments %>'
+			AND post_status = 'publish'";
+
+		$results = $wpdb->get_results( $query ); // phpcs:ignore
+
+		$pages_list = array();
+
+		if ( $results ) {
+			foreach ( $results as $result ) {
+				$page_id       = $result->ID;
+				$get_permalink = get_permalink( $page_id );
+				if ( false !== $get_permalink && ! in_array( $page_id, $this->cleared_extra_items, true ) ) {
+					$pages_list[]                = $get_permalink;
+					$this->cleared_extra_items[] = $page_id;
+				}
+			}
+		}
+
+		if ( ! empty( $pages_list ) ) {
+			// CLear Cloudflare, if enabled.
+			Breeze_CloudFlare_Helper::purge_cloudflare_cache_urls( $pages_list );
+
+			// Remove local cache file.
+			foreach ( $pages_list as $url_path ) {
+				$this->clear_local_cache_for_urls( array( $url_path ) );
+
+				$main     = new Breeze_PurgeVarnish();
+				$item_url = untrailingslashit( $url_path ) . '/?breeze';
+				$main->purge_cache( $item_url );
+			}
+		}
+
 	}
 
 	/**
@@ -67,7 +121,7 @@ class Breeze_PurgeCache {
 		//delete minify
 		Breeze_MinificationCache::clear_minification();
 		//clear normal cache
-		Breeze_PurgeCache::breeze_cache_flush();
+		Breeze_PurgeCache::breeze_cache_flush( true, true, true );
 		//do_action( 'breeze_clear_all_cache' );
 	}
 
@@ -91,6 +145,7 @@ class Breeze_PurgeCache {
 
 	//    Automatically purge all file based page cache on post changes
 	public function purge_post_on_update( $post_id ) {
+
 		$post_type = get_post_type( $post_id );
 
 		if ( ( defined( 'DOING_AUTOSAVE' ) && DOING_AUTOSAVE ) || 'revision' === $post_type ) {
@@ -104,7 +159,8 @@ class Breeze_PurgeCache {
 			$do_cache_reset = false;
 		}
 		$clear_wp_cache = true;
-		if ( defined( 'RedisCachePro\Version' ) ) {
+
+		if ( true === self::is_pro_plugin_ob_cache_enabled() ) {
 			$clear_wp_cache = false;
 		}
 
@@ -118,18 +174,30 @@ class Breeze_PurgeCache {
 		}
 	}
 
-	//    Automatically purge all file based term cache on post changes
-	public function purge_term_on_update() {
+	/**
+	 * Purges cache related to a term when it is updated.
+	 *
+	 * This method handles clearing object cache and file-based caching mechanisms
+	 * for the specified term. It ensures cache consistency across term updates
+	 * based on plugin configuration and active caching mechanisms.
+	 *
+	 * @param int $term_id The ID of the term being updated.
+	 * @param int $tt_id The term taxonomy ID associated with the term.
+	 * @param string $taxonomy The taxonomy name of the term being updated.
+	 *
+	 * @return void
+	 */
+	public function purge_term_on_update( int $term_id, int $tt_id, string $taxonomy ) {
 
-		$do_cache_reset = true;
 		$clear_wp_cache = true;
-		if ( defined( 'RedisCachePro\Version' ) ) {
+		if ( true === self::is_pro_plugin_ob_cache_enabled() ) {
 			$clear_wp_cache = false;
 		}
 
 		// File based caching only
 		if ( ! empty( Breeze_Options_Reader::get_option_value( 'breeze-active' ) ) ) {
-			self::breeze_cache_flush( $do_cache_reset, $clear_wp_cache );
+			self::clear_op_cache_for_terms( $term_id, $tt_id, $taxonomy );
+			self::breeze_cache_flush( true, $clear_wp_cache );
 		}
 	}
 
@@ -155,8 +223,14 @@ class Breeze_PurgeCache {
 				return;
 			}
 
-			$this->purge_cloudflare_cache( $post_id );
+			$list_of_urls = $this->collect_urls_for_cache_purge( $post_id );
 
+			if ( ! empty( $list_of_urls ) ) {
+				// Purge local cache for the URLs list.
+				$this->clear_local_cache_for_urls( $list_of_urls );
+				// Purge CF cache.
+				Breeze_CloudFlare_Helper::purge_cloudflare_cache_urls( $list_of_urls );
+			}
 		}
 
 	}
@@ -169,13 +243,21 @@ class Breeze_PurgeCache {
 	 * @return void
 	 */
 	public function purge_post_on_trash( int $post_id ) {
-		$this->purge_cloudflare_cache( $post_id );
+		$list_of_urls = $this->collect_urls_for_cache_purge( $post_id );
+
+		if ( ! empty( $list_of_urls ) ) {
+			// Purge local cache for the URLs list.
+			$this->clear_local_cache_for_urls( $list_of_urls );
+			// Purge CF cache.
+			Breeze_CloudFlare_Helper::purge_cloudflare_cache_urls( $list_of_urls );
+		}
+
 	}
 
-	private function purge_cloudflare_cache( $post_id ) {
+	private function collect_urls_for_cache_purge( $post_id ): array {
 
 		if ( false === get_permalink( $post_id ) ) {
-			return;
+			return array();
 		}
 		// Reset CloudFlare cache.
 		$get_permalink = get_permalink( $post_id );
@@ -224,11 +306,18 @@ class Breeze_PurgeCache {
 		$categories = get_the_category( $post_id );
 		if ( $categories ) {
 			foreach ( $categories as $cat ) {
+				$category_link        = get_category_link( $cat->term_id );
+				$category_link_no_cat = str_replace( 'category/', '', $category_link );
+				if ( ! empty( $category_link ) && $category_link !== $category_link_no_cat ) {
+					array_push( $list_of_urls, $category_link_no_cat );
+				}
+
 				array_push(
 					$list_of_urls,
-					get_category_link( $cat->term_id ),
+					$category_link,
 					get_rest_url() . $rest_api_route . '/categories/' . $cat->term_id . '/'
 				);
+				$category_link = '';
 			}
 		}
 
@@ -245,7 +334,8 @@ class Breeze_PurgeCache {
 
 		// Archives and their feeds
 		if ( $this_post_type && ! in_array( $this_post_type, $noarchive_post_type, true ) ) {
-			$get_archive_link      = get_post_type_archive_link( get_post_type( $post_id ) );
+			$get_archive_link = get_post_type_archive_link( get_post_type( $post_id ) );
+
 			$get_archive_feed_link = get_post_type_archive_feed_link( get_post_type( $post_id ) );
 			if ( ! empty( $get_archive_link ) ) {
 				$list_of_urls[] = $get_archive_link;
@@ -269,7 +359,6 @@ class Breeze_PurgeCache {
 		array_push(
 			$list_of_urls,
 			get_rest_url(),
-			trailingslashit( home_url() )
 		);
 		if ( 'page' === get_option( 'show_on_front' ) ) {
 			// Ensure we have a page_for_posts setting to avoid empty URL
@@ -278,11 +367,58 @@ class Breeze_PurgeCache {
 			}
 		}
 
-		Breeze_CloudFlare_Helper::purge_cloudflare_cache_urls( $list_of_urls );
+		// Trim all URLs in the list to ensure clean output
+		$list_of_urls = array_map( 'trim', $list_of_urls );
+
+		$homepage = trailingslashit( home_url() );
+		if ( ! in_array( $homepage, $list_of_urls, true ) ) {
+			// Clear the cache for homepage
+			array_push(
+				$list_of_urls,
+				$homepage
+			);
+		}
+
+		return $list_of_urls;
 	}
 
+	/**
+	 * Clears the local cache for the specified URLs.
+	 *
+	 * @param array $list_of_urls An array of URLs for which the local cache should be cleared.
+	 *
+	 * @return void
+	 */
+	private function clear_local_cache_for_urls( array $list_of_urls ) {
+		global $wp_filesystem;
+		if ( empty( $wp_filesystem ) ) {
+			require_once ABSPATH . '/wp-admin/includes/file.php';
+			WP_Filesystem();
+		}
+
+		if ( ! empty( $list_of_urls ) ) {
+			foreach ( $list_of_urls as $local_url ) {
+				if ( $wp_filesystem->exists( breeze_get_cache_base_path() . hash( 'sha512', $local_url ) ) ) {
+					$wp_filesystem->rmdir( breeze_get_cache_base_path() . hash( 'sha512', $local_url ), true );
+				}
+			}
+		}
+	}
+
+	/**
+	 * Purges the cache for a post associated with a newly approved comment.
+	 *
+	 * This method is triggered when a new comment is added and approved, handling cache
+	 * clearing for the related post using file-based caching and Cloudflare integration.
+	 *
+	 * @param int $comment_ID The ID of the new comment.
+	 * @param int|string $approved Approval status of the comment. Non-zero value indicates approval.
+	 * @param array $commentdata An array of comment data including the ID of the associated post.
+	 *
+	 * @return void
+	 */
 	public function purge_post_on_new_comment( $comment_ID, $approved, $commentdata ) {
-		if ( empty( $approved ) ) {
+		if ( 1 !== $approved ) {
 			return;
 		}
 		// File based caching only
@@ -291,40 +427,40 @@ class Breeze_PurgeCache {
 
 			Breeze_CloudFlare_Helper::purge_cloudflare_cache_urls( array( get_permalink( $post_id ) ) );
 
-			global $wp_filesystem;
-
-			if ( empty( $wp_filesystem ) ) {
-				require_once( ABSPATH . '/wp-admin/includes/file.php' );
-				WP_Filesystem();
-			}
-
 			$url_path = get_permalink( $post_id );
-			if ( $wp_filesystem->exists( breeze_get_cache_base_path() . hash( 'sha512', $url_path ) ) ) {
-				$wp_filesystem->rmdir( breeze_get_cache_base_path() . hash( 'sha512', $url_path ), true );
-			}
+			// Purge local cache for the URLs list.
+			$this->clear_local_cache_for_urls( array( $url_path ) );
+
+			$this->detect_comments_page_clear_cache();
+			$this->clear_op_cache_for_comments( $comment_ID );
 		}
 	}
 
-	//            if a comments status changes, purge it's parent posts cache
-	public function purge_post_on_comment_status_change( $comment_ID, $comment_status ) {
+	/**
+	 * Purges the post associated with a comment when the comment's status changes.
+	 *
+	 * This method is triggered to handle cache clearing for posts related to comments,
+	 * based on file-based caching settings and configurations specific to Breeze.
+	 *
+	 * @param int $comment_ID The ID of the comment whose status has changed.
+	 * @param mixed $comment_status The new status of the comment.
+	 *
+	 * @return void
+	 */
+	public function purge_post_on_comment_status_change( int $comment_ID, $comment_status ) {
 		// File based caching only
 		if ( ! empty( Breeze_Options_Reader::get_option_value( 'breeze-active' ) ) ) {
 			$comment = get_comment( $comment_ID );
 			if ( ! empty( $comment ) ) {
 				$post_id = $comment->comment_post_ID;
 
-				global $wp_filesystem;
-
-				WP_Filesystem();
-
 				$url_path = get_permalink( $post_id );
-
 				Breeze_CloudFlare_Helper::purge_cloudflare_cache_urls( array( $url_path ) );
-
-				if ( $wp_filesystem->exists( breeze_get_cache_base_path() . hash( 'sha512', $url_path ) ) ) {
-					$wp_filesystem->rmdir( breeze_get_cache_base_path() . hash( 'sha512', $url_path ), true );
-				}
+				$this->clear_local_cache_for_urls( array( $url_path ) );
 			}
+
+			$this->detect_comments_page_clear_cache();
+			$this->clear_op_cache_for_comments( $comment_ID );
 		}
 	}
 
@@ -336,17 +472,29 @@ class Breeze_PurgeCache {
 	 *
 	 * @return void
 	 */
-	public static function breeze_cache_flush( $flush_cache = true, $clear_ocp = true ) {
-		global $wp_filesystem, $post;
+	public static function breeze_cache_flush( $flush_cache = true, $clear_ocp = true, $purge_all_html_folder = false ) {
+		global $post;
+
+		$post_id = null;
+		if ( $post instanceof WP_Post ) {
+			$post_id = $post->ID;
+		}
+
 		if ( true === Breeze_CloudFlare_Helper::is_log_enabled() ) {
 			error_log( '######### PURGE LOCAL CACHE HTML ###: ' . var_export( 'true', true ) );
 		}
-		require_once( ABSPATH . 'wp-admin/includes/file.php' );
 
-		WP_Filesystem();
+		if ( true === $purge_all_html_folder ) {
+			global $wp_filesystem;
 
-		$cache_path = breeze_get_cache_base_path( is_network_admin() );
-		$wp_filesystem->rmdir( untrailingslashit( $cache_path ), true );
+			if ( empty( $wp_filesystem ) ) {
+				require_once ABSPATH . '/wp-admin/includes/file.php';
+				WP_Filesystem();
+			}
+
+			$cache_path = breeze_get_cache_base_path( is_network_admin() );
+			$wp_filesystem->rmdir( untrailingslashit( $cache_path ), true );
+		}
 
 		if ( true === $flush_cache && ! empty( $post ) && is_object( $post ) ) {
 			$post_type = get_post_type( $post->ID );
@@ -356,7 +504,7 @@ class Breeze_PurgeCache {
 				'tribe_events',
 				'shop_order',
 			);
-			if ( in_array( $post_type, $ignore_object_cache ) ) {
+			if ( in_array( $post_type, $ignore_object_cache, true ) ) {
 				$flush_cache = false;
 			}
 		}
@@ -365,14 +513,11 @@ class Breeze_PurgeCache {
 			$flush_cache = false;
 		}
 
-		if ( function_exists( 'wp_cache_flush' ) && true === $flush_cache && true === $clear_ocp ) {
+		if ( ! empty( $post_id ) && true === $flush_cache && true === $clear_ocp ) {
 			if ( true === Breeze_CloudFlare_Helper::is_log_enabled() ) {
 				error_log( '######### PURGE OBJECT CACHE ###: ' . var_export( 'true', true ) );
 			}
-			#if ( ! defined( 'RedisCachePro\Version' ) && ! defined( 'WP_REDIS_VERSION' ) ) {
-			wp_cache_flush();
-			#}
-
+			self::clear_op_cache_for_posts( $post_id );
 		}
 	}
 
@@ -416,6 +561,16 @@ class Breeze_PurgeCache {
 	}
 
 
+	/**
+	 * Flushes the entire object cache for the current context.
+	 *
+	 * This method resets the object cache, clearing all stored data. When executed
+	 * on a network admin screen, it will handle any additional logic specific to
+	 * network context (if required). Otherwise, it flushes the cache for the current
+	 * instance or site.
+	 *
+	 * @return bool True on success, false on failure.
+	 */
 	public static function __flush_object_cache() {
 		set_as_network_screen();
 
@@ -425,6 +580,107 @@ class Breeze_PurgeCache {
 		}
 
 		return wp_cache_flush();
+	}
+
+	/**
+	 * Clears the object cache associated with a specific post or custom post type.
+	 *
+	 * This method removes cached data for the given post ID, including post metadata
+	 * and any related term cache associated with the post.
+	 *
+	 * @param int $object_id The ID of the post or custom post type for which the cache should be cleared.
+	 *
+	 * @return void
+	 */
+	public static function clear_op_cache_for_posts( int $object_id ) {
+		if ( true === self::is_pro_plugin_ob_cache_enabled() ) {
+			return;
+		}
+
+		// POST/PAGE/CPT
+		// $object_id is $post_id
+		wp_cache_delete( $object_id, 'posts' );
+		wp_cache_delete( $object_id, 'post_meta' ); // If needed
+		wp_cache_delete( 'post_terms_' . $object_id, 'terms' ); //for term related to post
+
+	}
+
+	/**
+	 * Clears the object cache related to specific terms and their taxonomy.
+	 *
+	 * This method removes cached data for the given term, term taxonomy, and associated taxonomy
+	 * data, ensuring that outdated cached information is invalidated and updated.
+	 *
+	 * @param int $object_id The ID of the term for which the cache should be cleared.
+	 * @param int $tt_id The term taxonomy ID associated with the term.
+	 * @param string $taxonomy The taxonomy name related to the term.
+	 *
+	 * @return void
+	 */
+	private static function clear_op_cache_for_terms( int $object_id, int $tt_id, string $taxonomy ) {
+		if ( true === self::is_pro_plugin_ob_cache_enabled() ) {
+			return;
+		}
+
+		// TAXONOMY
+		// $object_id is term_id
+		wp_cache_delete( $object_id, 'terms' );
+		wp_cache_delete( $tt_id, 'term_taxonomy' );
+		#delete_transient( 'all_terms' ); //for all terms, if needed
+		$tax_cache_key = "{$taxonomy}_terms"; //dynamic based on current taxonomy
+		wp_cache_delete( $tax_cache_key, 'terms' ); //clear cache for taxonomy
+
+	}
+
+	/**
+	 * Clears the object cache related to a specific comment and its associated post.
+	 *
+	 * This method removes cached data for the given comment, resets queries related
+	 * to comments, clears the comment count cache, and clears the cache for the post
+	 * associated with the given comment.
+	 *
+	 * @param int $comment_id The ID of the comment for which the cache should be cleared.
+	 *
+	 * @return void
+	 */
+	private static function clear_op_cache_for_comments( int $comment_id ) {
+		if ( true === self::is_pro_plugin_ob_cache_enabled() ) {
+			return;
+		}
+		error_log( '$comment_id: ' . var_export( $comment_id, true ) );
+		wp_cache_delete( $comment_id, 'comment' );
+		wp_cache_delete( 'comment_query_1_' . get_the_ID(), 'comment' );
+		wp_cache_delete( 'get_comments_by_post_id_' . get_the_ID(), 'comment' );
+		//Clear all comments count cache, as it has changed
+		wp_cache_delete( 'comments-per-page', 'counts' );
+
+		// Clear the cache for the post the comment belongs to
+		$post_id = get_comment( $comment_id )->comment_post_ID;
+		wp_cache_delete( $post_id, 'posts' );
+		wp_cache_delete( $post_id, 'post_meta' ); // If needed
+	}
+
+	/**
+	 * Checks if object cache support is enabled by a Pro plugin.
+	 *
+	 * This method verifies the presence of object caching functionality typically provided
+	 * by Pro-level caching plugins, such as Redis Cache Pro or similar plugins.
+	 *
+	 * @return bool Returns true if a Pro plugin enabling object cache is detected, otherwise false.
+	 */
+	private static function is_pro_plugin_ob_cache_enabled(): bool {
+		// If Redis Pro is found, leave the object cache clear to it.
+		if ( defined( 'RedisCachePro\Version' ) || defined( 'WP_REDIS_VERSION' ) ) {
+			return true;
+		}
+
+		// if the function does not exist, block the calling of it.
+		if ( ! function_exists( 'wp_cache_delete' ) ) {
+			return true;
+		}
+		// Go ahead and clear the object cache.
+		return false;
+
 	}
 
 }
